Recommend gir1.2-appindicator3-0.1 for appindicator support.
As seen in virtManager/systray.py appindicator is an optional runtime
dependency which then would provide an indicator as little helper.
Adding a dependency to gir1.2-appindicator3-0.1 will make that work
reliably.

It is highly useful and I think the default user of virt-manager is UI
driven and would appreciate (more than a suggests), OTOH it isn't strictly
needed to be able to work (no depends) - therefore I choose "Recommends".

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>